+++ src/main/java/lotto/domain/summary/Summary.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.domain.summary;
 
 import lotto.constants.Winning;
